uchar: Work-around qemu bug on arm
Building either of these configurations using -fsanitize=undefined:

arm-none-eabi/armv8.1m.main_hard_nofp_mve_pacret_bti_exn_rtti_unaligned_size
arm-none-eabi/armv8.1m.main_hard_nofp_mve_pacret_bti_unaligned_size

causes current tcg to mis-transpile this code from mbrtoc8.c

        while (c32 >= ((char32_t)1 << (6 * count + (6 - count)))) {
	}

and generate a false mul_overflow exception. Work around this by
switching to unsigned arithmetic, which is harmless as the values are
always positive.

+++ libc/uchar/mbrtoc8.c
@@ -44,12 +44,12 @@ mbrtoc8(char8_t * __restrict pc8, const char * __restrict s, size_t n, mbstate_t
         ps = &mbrtoc8_state;
 
     if (ps->__count < 0) {
-        int count = -ps->__count;
+        unsigned count = -ps->__count;
 
         count--;
         if (pc8 != NULL)
             *pc8 = 0x80 | ((ps->__value.__ucs >> (6 * count)) & 0x3f);
-        ps->__count = -count;
+        ps->__count = -(int)count;
         return (size_t)-3;
     }
 
@@ -66,8 +66,8 @@ mbrtoc8(char8_t * __restrict pc8, const char * __restrict s, size_t n, mbstate_t
     }
 
     if (!char32_is_one_byte(c32)) {
-        int     count = 1;
-        char8_t c8;
+        unsigned count = 1;
+        char8_t  c8;
 
         c8 = 0xc0;
         while (c32 >= ((char32_t)1 << (6 * count + (6 - count)))) {
@@ -75,7 +75,7 @@ mbrtoc8(char8_t * __restrict pc8, const char * __restrict s, size_t n, mbstate_t
             count++;
         }
         c8 |= (c32 >> (6 * count));
-        ps->__count = -count;
+        ps->__count = -(int)count;
         ps->__value.__ucs = c32;
         if (pc8 != NULL)
             *pc8 = c8;
